third_party/WebKit/LayoutTests/http/tests/security/link-crossorigin-stylesheet-no-cors-expected.txt

CONSOLE ERROR: Access to CSS stylesheet at 'http://localhost:8080/security/resources/green-background.css?1' from origin 'http://localhost:8080' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to CSS stylesheet at 'http://localhost:8080/security/resources/green-background.css?2' from origin 'http://localhost:8080' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to CSS stylesheet at 'http://localhost:8080/security/resources/green-background.css?8' from origin 'http://localhost:8080' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to CSS stylesheet at 'http://localhost:8080/security/resources/green-background.css?9' from origin 'http://localhost:8080' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
PASS 
PASS 
PASS 
PASS 
PASS 
PASS 
PASS 
PASS 
PASS successfullyParsed is true

TEST COMPLETE
Test that a linked stylesheet with a crossorigin attribute does not load a cross-origin resource that isn't CORS enabled.
